refactor to add a createScheduleEvent effect
